+++ app/exceptions/device_exception.py
@@ -1,5 +1,9 @@
-class DeviceAlreadyRegisteredError(Exception):
-    pass
+from app.exceptions.base_exception import NotFoundError, AlreadyExistsError
 
-class DeviceNotFoundError(Exception):
-    pass
+class DeviceAlreadyRegisteredError(AlreadyExistsError):
+    def __init__(self, device_id: str):
+        super().__init__("Device", device_id)
+
+class DeviceNotFoundError(NotFoundError):
+    def __init__(self, device_id: str):
+        super().__init__("Device", device_id)