+++ app/repositories/device_repository.py
@@ -38,3 +38,15 @@ class DeviceRepository:
             select(Device).where(Device.device_id == device_id)
         )
         return result.scalar_one_or_none()
+    
+    #Updates the spotify_device_id of a device
+    async def update_heartbeat(self, device_id: str, spotify_device_id: str) -> Device:
+        result  = await self.db.execute(
+            select(Device).where(Device.device_id == device_id)
+        )
+        device = result.scalar_one_or_none()
+        device.spotify_device_id = spotify_device_id
+        device.last_seen = datetime.utcnow()
+        await self.db.commit()
+        await self.db.refresh(device)
+        return device