+++ app/repositories/user_repository.py
@@ -48,3 +48,15 @@ class UserRepository:
             select(User).where(User.id == user_id)
         )
         return result.scalar_one_or_none()
+    
+    #Updates the access token of a user, returns the status of the operation
+    async def update_tokens(self,user_id: int, access_token: str, token_expires_at: datetime) -> User:
+        result = await self.db.execute(
+            select(User).where(User.id == user_id)
+        )
+        user = result.scalar_one_or_none()
+        user.access_token = access_token
+        user.token_expires_at = token_expires_at
+        await self.db.commit()
+        await self.db.refresh(user)
+        return user